refacrtor: replace theme switcher to separate component
