+++ src/client/ui/header/header.tsx
@@ -1,17 +1,8 @@
 import React from 'https://esm.sh/react@18.2.0'
 import { API_URL } from '../../../const.ts'
-import { Select } from '../select/select.tsx'
+import { Theme } from '../theme/theme.tsx'
 
 export const Header = () => {
-  const [theme, setTheme] = React.useState('dark')
-
-  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const { checked } = event.target
-    const newTheme = checked ? 'dark' : 'light'
-    setTheme(newTheme)
-    document.documentElement.className = `theme-${newTheme}`
-  }
-
   return (
     <header className="header">
       <a href="/" className="header__title">
@@ -20,10 +11,7 @@ export const Header = () => {
         </div>
         <div> Badge service</div>
       </a>
-      <label className="header__theme">
-        <input type="checkbox" checked={theme === 'dark'} onChange={handleChange} />
-        <img src={`${API_URL}/img/${theme}`} alt="theme" />
-      </label>
+      <Theme />
     </header>
   )
 }
