+++ src/client/ui/theme/theme.tsx
@@ -1,0 +1,20 @@
+import React from 'https://esm.sh/react@18.2.0'
+import { API_URL } from '../../../const.ts'
+
+export const Theme = () => {
+  const [theme, setTheme] = React.useState('dark')
+
+  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const { checked } = event.target
+    const newTheme = checked ? 'dark' : 'light'
+    setTheme(newTheme)
+    document.documentElement.className = `theme-${newTheme}`
+  }
+
+  return (
+    <label className="header__theme">
+      <input type="checkbox" checked={theme === 'dark'} onChange={handleChange} />
+      <img src={`${API_URL}/img/${theme}`} alt="theme" />
+    </label>
+  )
+}
